# hops-factory/supplier-service-module/supplier-service-client/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: Supplier Service
  version: 1.0.1-SNAPSHOT
basePath: /supplier_service
tags:
  - name: healthz
  - name: supplier
paths:
  /healthz:
    get:
      tags:
        - healthz
      summary: Perform healt hcheck operation.
      description: ''
      operationId: basicHealthcheck
      consumes:
        - text/plain
      produces:
        - text/plain
      parameters: []
      responses:
        '200':
          description: OK
        '503':
          description: Health check failed.
  /supplier:
    get:
      tags:
        - supplier
      summary: Find all suppliers
      description: ''
      operationId: getAllSuppliers
      consumes:
        - 'application/json, application/supplier.service.v1+json'
      produces:
        - 'application/json, application/supplier.service.v1+json'
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            type: array
            items:
              $ref: '#/definitions/Supplier'
        '500':
          description: Internal server error.
    post:
      tags:
        - supplier
      summary: Create new supplier
      description: ''
      operationId: insertNewSupplier
      consumes:
        - 'application/json, application/supplier.service.v1+json'
      produces:
        - 'application/json, application/supplier.service.v1+json'
      parameters:
        - in: body
          name: body
          description: Supplier to be created
          required: true
          schema:
            $ref: '#/definitions/Supplier'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Supplier'
        '400':
          description: Invalid request
  '/supplier/{id}':
    get:
      tags:
        - supplier
      summary: Find supplier with given id.
      description: ''
      operationId: getSupplierById
      consumes:
        - 'application/json, application/supplier.service.v1+json'
      produces:
        - 'application/json, application/supplier.service.v1+json'
      parameters:
        - name: id
          in: path
          description: Id of the requested supplier
          required: true
          type: string
          format: uuid
          enum:
            - Valid UUID strings
          x-example: 571fbefb-f96e-40c7-b699-94ac2403eab4
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Supplier'
        '404':
          description: Request with given id does not exist.
        '500':
          description: Internal server error.
definitions:
  Supplier:
    type: object
    properties:
      id:
        type: string
        example: 571fbefb-f96e-40c7-b699-94ac2403eab4
        description: 'Id of this supplier, in form of uuid as string'
      name:
        type: string
        description: Name of this supplier
      address:
        type: string
        description: Address of this supplier
      phone:
        type: string
        description: 'Phone number of this supplier, as string, no specific format.'
      email:
        type: string
        example: someone@somewhere.com
        description: Email address of this supplier as string.
    description: Supplier class